Market_survey/requirements.txt
langchain<0.2
langchain-openai>=0.1.0
python-dotenv>=1.0.0
